Refactor Unicode feature
