+++ quantum/process_keycode/process_unicode.h
@@ -16,6 +16,9 @@
 
 #pragma once
 
-#include "process_unicode_common.h"
+#include <stdbool.h>
+#include <stdint.h>
+
+#include "action.h"
 
 bool process_unicode(uint16_t keycode, keyrecord_t *record);
